requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Data processing
openpyxl>=3.1.0
PyPDF2>=3.0.0
python-multipart>=0.0.6
chardet>=5.0.0

# SSH and terminal access
paramiko>=3.2.0
psutil>=5.9.0
scp>=0.14.0

# Machine learning
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Vector storage (local only)
chromadb>=0.4.0

# Basic HTTP (no external APIs)
requests>=2.31.0

# GUI dependencies
streamlit>=1.25.0
plotly>=5.15.0
altair>=5.0.0

# CLI dependencies
click>=8.1.0
rich>=13.0.0
typer>=0.9.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
jinja2>=3.1.0
markdown>=3.4.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0